+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Reorder stats (tables at the top)
 * Check bookings data fields length before submitting
+* Booking Form: Allow admins to not fill email field
 
 ### 1.8.5: October 21st, 2021
 * Admins can book at a later date